+++ src/main/java/techcourse/myblog/web/core/AuthWebMvcConfig.java
@@ -7,6 +7,8 @@ import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import techcourse.myblog.web.interceptor.AuthInterceptor;
+
 @Configuration
 public class AuthWebMvcConfig implements WebMvcConfigurer {
