+++ src/main/java/techcourse/myblog/web/core/CertificationWebMvcConfig.java
@@ -3,28 +3,27 @@ package techcourse.myblog.web.core;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurationSupport;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import java.util.List;
+import techcourse.myblog.web.interceptor.CertificationInterceptor;
 
 @Configuration
-public class MyPageWebMvcConfig implements WebMvcConfigurer {
+public class CertificationWebMvcConfig implements WebMvcConfigurer {
 
-    @Qualifier(value = "myPageInterceptor")
+    @Qualifier(value = "certificationInterceptor")
     private HandlerInterceptor handlerInterceptor;
 
     @Autowired
-    public MyPageWebMvcConfig(MyPageInterceptor myPageInterceptor) {
-        this.handlerInterceptor = myPageInterceptor;
+    public CertificationWebMvcConfig(CertificationInterceptor certificationInterceptor) {
+        this.handlerInterceptor = certificationInterceptor;
     }
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(handlerInterceptor)
-                .addPathPatterns("/mypage/**");
+                .addPathPatterns("/mypage/**")
+                .addPathPatterns("/logout");
     }
 }
